Move interfaces with contracts from dist to test
