+++ universal-login-contracts/test/contracts/Wallet/WalletProxy.test.ts
@@ -5,7 +5,7 @@ import {Contract, Wallet} from 'ethers';
 import {DEFAULT_GAS_PRICE, DEFAULT_GAS_LIMIT} from '@universal-login/commons';
 import WalletProxy from '../../../dist/contracts/WalletProxy.json';
 import basicWalletAndProxy from '../../fixtures/basicWalletAndProxy';
-import {MockWalletMasterInterface} from '../../../src/interfaces';
+import {MockWalletMasterInterface} from '../../helpers/interfaces';
 
 chai.use(chaiAsPromised);
 chai.use(solidity);
